backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-dotenv
python-jose[cryptography]
bcrypt
pydantic
pydantic-settings
cryptography
python-multipart
openpyxl
pandas

